move files decoder to kernel
