+++ src/main/java/Kernel/decoder/Decoder.java
@@ -1,4 +1,4 @@
-package decoder.core;
+package  Kernel.decoder;
 
 import java.text.Normalizer;
 import java.util.*;
