+++ src/main/java/Kernel/decoder/EncoderMultipleOrders.java
@@ -1,4 +1,4 @@
-package decoder.core;
+package Kernel.decoder;
 
 import java.util.List;
